Build fix for Buildbot.
        * DumpRenderTree/DumpRenderTree.mm:
        (dump):



Canonical link: https://commits.webkit.org/19987@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@25517 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebKitTools/DumpRenderTree/DumpRenderTree.mm
@@ -911,7 +911,7 @@ void dump(void)
             WebView* view = [mainFrame webView];
             NSSize webViewSize = [view frame].size;
 
-            CGContextRef cgContext = CGBitmapContextCreate(screenCaptureBuffer, webViewSize.width, webViewSize.height, 8, webViewSize.width * 4, sharedColorSpace, kCGBitmapByteOrder32Host | kCGImageAlphaPremultipliedLast);
+            CGContextRef cgContext = CGBitmapContextCreate(screenCaptureBuffer, static_cast<size_t>(webViewSize.width), static_cast<size_t>(webViewSize.height), 8, static_cast<size_t>(webViewSize.width) * 4, sharedColorSpace, kCGBitmapByteOrder32Host | kCGImageAlphaPremultipliedLast);
 
             NSGraphicsContext* savedContext = [[[NSGraphicsContext currentContext] retain] autorelease];
             NSGraphicsContext* nsContext = [NSGraphicsContext graphicsContextWithGraphicsPort:cgContext flipped:NO];
